Modifed the mapping to use simple array instead of Hashtable look-ups. The Hashtable looks-ups are O(1) but take 100s of CPU-instructions (to generate hash) versus array look-ups are O(1) but takes 2-or-3 CPU-instructions, making them much-much-faster.  With this change the simulations run a bit faster.
It is important to note that Big-O is not really useful in practice --
it is just to give some idea about algorithms from a theoretical
perspective.  In practice, you should not rely on Big-O for anything
meaningful.  That is why most of the API documentations does not even
bother to record Big-O because it is very misleading.