A couple of minor changes to Antigen.java to correctly process command-line arguments for pEpConv and the mutation file.
Minor changes to the GSA scripts to account for the fact that envrionmental
aspects (which are very important) are currently not being used by HA-Seq 
testing. 

Added an R script to plot correlations between parameters based on 
successful (i.e. average number of clusters is approximately equal to the
expected number of clusters) combinations.