Changes to the web-UI and associated helper/controller classes to provide two key integration points with IBIS (https://ibis.eha.io) --
1. The first set of features enable displaying data about "Airport riks" and
   "Outbound traffic" from IBIS in the popup window.

2. The second integration permits IBIS to redirect a user to Foresight to 
   setup an analysis with airport infection pre-populated.

There are several changes required to support these features. Unfortunately,
I missed checking-in the changes as they were happening. So this is a bit of
a bulk update. Please refer to the detailed comments in each file to track
the changes to the individual files for the above 2 major functionalities.

These are the initial changes to the features. Some more changes may be needed
as we go through more detailed testing.