Added feature to create and manage new "kind" of custom models, in addition to the default 'chikv' and 'denv' types.  This just builds-upon/extends the custom model generation capability. Implementing this feature required the following changes:
ModelController.php: Custom model generation now permits
a "destKind" (destination kind) string to be specified by the
user.  This destination kind string is used to create a new
'kind' of models.  This also causes a new sub-directory to be
created in "storage/app/analysis/scripts" and the
needed "calib_params.csv" and "param_overrides.sh" file are
copied from the source 'kind'.

Model.php: A couple of convenience functions to check if a given
'kind' string is valid (i.e., present in the 'models' table) and
to obtain a list of valid model kinds.  This list is used in the
dashboard for displaying the correct number of tabs.

dashboard.blade.php: Now the dashboard uses a list of model
'kind's from the database rather than the 2 fixed 'chikv' and
'zikv' kinds.  However, there are are checks to ensure that those
2 kinds continue to get the same display name thereby maintaining
consistency with the current display.

webui/resources/views/generate.blade.php: Permits the user to
change the 'kind' value associated with a custom model being
generated.