Added a feature that permits an instructor to download preloaded files. Now, each of the preloaded files is a link that an instructor can click. Upon clicking the link, the file is downloaded.
The logic for downloading the file is in InstructorController.php.  The links
are generated in UploadedFile.blade.php which is included in
AssignmentSettings.blade.php.  The assignment settings had to be changed
from an @each() to a @foreach() blade directive to ease passing variables
to the uploaded file blade.

Added a new route to web.php to tie the URLs generated by the blades to
the logic in InstructorController.

Tested the links and they seem to be working fine.