Committing several changes and enhancement to cityscape. This commit includes changes for doing a radius filter. The biggest change is that now cityscape uses MPI for accelerating model generation using multiple compute nodes on the cluster.  In addition, the changes include a centralized load balancing strategy in which each thread picks up the next building to be processed from a centralized counter. The centralized counter is implemented using MPI 5.0's window feature along with an atomic fetch_and_op feature. So now cityscape requires MPI 5.0 or newer to compile and run.
Now, the building assignment is stored in a schedule object -- see
ScheduleEntry for additional details.  This aspect of the system is
still work in progress and needs a lot more features added before it is
ready for actual testing/use.

Also changed PathFinderTester and PathFinder to track and report
additional information (such as: both distance and time) along with the
haversine distance between the start and end buildings.